+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Unreleased
+    - Added transform column method to dataset object
     - Added describe method to dataset object
     - Added deduplicate method to dataset object
     - Added unzip static factory for Labeled datasets from data table